move SlotTagModel type to SlotTag class
