+++ packages/app/src/classes/Tags/KeeperTag.ts
@@ -1,4 +1,4 @@
-import { SlotTagModel } from '@cambrian/app/models/SlotTagModel'
+import { SlotTagModel } from '@cambrian/app/src/classes/Tags/SlotTag'
 import SlotTag from './SlotTag'
 
 export class KeeperTag extends SlotTag {
