+++ packages/app/src/components/info/solver/BaseSolverInfo.tsx
@@ -6,7 +6,7 @@ import ModalHeader from '../../layout/header/ModalHeader'
 import { OutcomeCollectionModel } from '@cambrian/app/models/OutcomeCollectionModel'
 import OutcomeOverview from '@cambrian/app/ui/solver/OutcomeOverview'
 import RecipientInfoItem from '../RecipientInfo'
-import { SlotTagsHashMapType } from '@cambrian/app/models/SlotTagModel'
+import { SlotTagsHashMapType } from '@cambrian/app/src/classes/Tags/SlotTag'
 import SolverConfigItem from '../../list/SolverConfigItem'
 import { SolverTagModel } from '@cambrian/app/models/SolverTagModel'
 import { TokenModel } from '@cambrian/app/models/TokenModel'
