+++ packages/app/src/ui/composer/controls/solver/slotList/forms/CreateSlotForm.tsx
@@ -10,7 +10,7 @@ import SelectSlotType from '@cambrian/app/components/selects/SelectSlotType'
 import SelectSolverFunction from '@cambrian/app/components/selects/SelectSolverFunction'
 import SlotDataInputField from '@cambrian/app/components/inputs/SlotDataInput'
 import SlotTagFormFields from '../../general/forms/SlotTagFormFields'
-import { SlotTagModel } from '@cambrian/app/models/SlotTagModel'
+import { SlotTagModel } from '@cambrian/app/src/classes/Tags/SlotTag'
 import { SlotType } from '@cambrian/app/models/SlotType'
 import { SolidityDataTypes } from '@cambrian/app/models/SolidityDataTypes'
 import _uniqueId from 'lodash/uniqueId'
